requirements.txt
coverage>=3.7.1
coveralls>=0.5
django-timezone-field==2.0
